requirements.txt
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.8.3
click==8.2.1
cloudinary==1.44.1
dnspython==2.7.0
Flask==3.1.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-PyMongo==3.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pymongo==4.14.0
python-dotenv==1.1.1
six==1.17.0
urllib3==2.5.0
Werkzeug==3.1.3
